requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
cffi==1.12.3
cryptography==2.7
Django==2.2.2
django-rest-knox==4.1.0
djangorestframework==3.9.4
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
gunicorn==19.7.1
whitenoise==3.3.1
